Move Tolstoy remote client into separate module.
