+++ tolstoy/src/lib.rs
@@ -33,6 +33,8 @@ extern crate mentat_core;
 extern crate rusqlite;
 extern crate uuid;
 
+mod remote_client;
+
 pub mod schema;
 pub mod metadata;
 pub mod tx_processor;
